+++ tests/mapper/test_node_mappers.py
@@ -20,7 +20,7 @@ Mapper tests:
 
 """
 from stellargraph.core.graph import *
-from stellargraph.mapper.node_mappers import *
+from stellargraph.mapper import *
 
 import networkx as nx
 import numpy as np
@@ -39,7 +39,7 @@ def example_graph_1(feature_size=None):
     # Add example features
     if feature_size is not None:
         for v in G.nodes():
-            G.node[v]["feature"] = np.ones(feature_size)
+            G.nodes[v]["feature"] = np.ones(feature_size)
         return StellarGraph(G, node_features="feature")
 
     else:
@@ -55,7 +55,7 @@ def example_graph_2(feature_size=None):
     # Add example features
     if feature_size is not None:
         for v in G.nodes():
-            G.node[v]["feature"] = int(v) * np.ones(feature_size, dtype="int")
+            G.nodes[v]["feature"] = int(v) * np.ones(feature_size, dtype="int")
         return StellarGraph(G, node_features="feature")
 
     else:
@@ -75,7 +75,7 @@ def example_graph_3(feature_size=None, n_edges=20, n_nodes=6, n_isolates=1):
     # Add example features
     if feature_size is not None:
         for v in G.nodes():
-            G.node[v]["feature"] = int(v) * np.ones(feature_size, dtype="int")
+            G.nodes[v]["feature"] = int(v) * np.ones(feature_size, dtype="int")
         return StellarGraph(G, node_features="feature")
 
     else:
@@ -90,7 +90,7 @@ def example_digraph_2(feature_size=None):
     # Add example features
     if feature_size is not None:
         for v in G.nodes():
-            G.node[v]["feature"] = np.ones(feature_size)
+            G.nodes[v]["feature"] = np.ones(feature_size)
         return StellarDiGraph(G, node_features="feature")
 
     else:
@@ -396,31 +396,6 @@ def test_nodemapper_isolated_nodes():
     assert pytest.approx(nf[2][2:]) == 0
 
 
-def test_nodemapper_manual_schema():
-    """
-    Tests checks on head nodes
-    """
-    n_feat = 4
-    n_batch = 2
-
-    # test graph
-    G = example_graph_1(feature_size=n_feat)
-
-    # Create manual schema
-    schema = G.create_graph_schema(create_type_maps=True)
-    GraphSAGENodeGenerator(G, schema=schema, batch_size=n_batch, num_samples=[1]).flow(
-        list(G)
-    )
-
-    # Create manual schema without type maps
-    # Currently this raises an error:
-    schema = G.create_graph_schema(create_type_maps=False)
-    with pytest.raises(RuntimeError):
-        GraphSAGENodeGenerator(
-            G, schema=schema, batch_size=n_batch, num_samples=[1]
-        ).flow(list(G))
-
-
 def test_nodemapper_incorrect_targets():
     """
     Tests checks on target shape
@@ -446,9 +421,11 @@ def test_hinnodemapper_constructor():
 
     # Should fail when head nodes are of different type
     with pytest.raises(ValueError):
-        HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2]).flow(G.nodes())
+        HinSAGENodeGenerator(
+            G, batch_size=2, num_samples=[2, 2], head_node_type="A"
+        ).flow(G.nodes())
 
-    gen = HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2])
+    gen = HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2], head_node_type="A")
     mapper = gen.flow([0, 1, 2, 3])
     assert gen.batch_size == 2
     assert mapper.data_size == 4
@@ -459,7 +436,7 @@ def test_hinnodemapper_constructor_all_options():
     feature_sizes = {"A": 10, "B": 10}
     G = example_hin_1(feature_sizes)
 
-    gen = HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2])
+    gen = HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2], head_node_type="A")
 
     nodes_of_type_a = G.nodes_of_type("A")
     mapper = gen.flow(nodes_of_type_a)
@@ -470,9 +447,9 @@ def test_hinnodemapper_constructor_all_options():
 def test_hinnodemapper_constructor_no_features():
     G = example_hin_1(feature_size_by_type=None)
     with pytest.raises(RuntimeError):
-        mapper = HinSAGENodeGenerator(G, batch_size=2, num_samples=[2, 2]).flow(
-            G.nodes()
-        )
+        mapper = HinSAGENodeGenerator(
+            G, batch_size=2, num_samples=[2, 2], head_node_type="A"
+        ).flow(G.nodes())
 
 
 def test_hinnodemapper_constructor_nx_graph():
@@ -489,9 +466,9 @@ def test_hinnodemapper_level_1():
     feature_sizes = {"t1": 1, "t2": 2}
     G, nodes_type_1, nodes_type_2 = example_hin_2(feature_sizes)
 
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[2]).flow(
-        nodes_type_2
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[2], head_node_type="t2"
+    ).flow(nodes_type_2)
 
     schema = G.create_graph_schema()
     sampling_adj = schema.type_adjacency_list(["t2"], 1)
@@ -515,9 +492,9 @@ def test_hinnodemapper_level_2():
     feature_sizes = {"t1": 1, "t2": 2}
     G, nodes_type_1, nodes_type_2 = example_hin_2(feature_sizes)
 
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[2, 3]).flow(
-        nodes_type_2
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[2, 3], head_node_type="t2"
+    ).flow(nodes_type_2)
 
     schema = G.create_graph_schema()
     sampling_adj = schema.type_adjacency_list(["t2"], 2)
@@ -547,9 +524,9 @@ def test_hinnodemapper_shuffle():
     feature_sizes = {"t1": 1, "t2": 4}
     G, nodes_type_1, nodes_type_2 = example_hin_2(feature_sizes)
 
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[0]).flow(
-        nodes_type_1, nodes_type_1, shuffle=True
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[0], head_node_type="t1"
+    ).flow(nodes_type_1, nodes_type_1, shuffle=True)
 
     expected_node_batches = [[3, 2], [1, 0]]
     assert len(mapper) == 2
@@ -568,9 +545,9 @@ def test_hinnodemapper_shuffle():
         assert all(np.array(nl) == expected_node_batches[ii])
 
     # With no shuffle
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[0]).flow(
-        nodes_type_1, nodes_type_1, shuffle=False
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[0], head_node_type="t1"
+    ).flow(nodes_type_1, nodes_type_1, shuffle=False)
     expected_node_batches = [[0, 1], [2, 3]]
     assert len(mapper) == 2
     for ii in range(len(mapper)):
@@ -586,9 +563,9 @@ def test_hinnodemapper_with_labels():
 
     labels = [n * 2 for n in nodes_type_1]
 
-    gen = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[2, 3]).flow(
-        nodes_type_1, labels, shuffle=False
-    )
+    gen = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[2, 3], head_node_type="t1"
+    ).flow(nodes_type_1, labels, shuffle=False)
     assert len(gen) == 2
 
     for ii in range(2):
@@ -618,17 +595,9 @@ def test_hinnodemapper_manual_schema():
 
     # Create manual schema
     schema = G.create_graph_schema(create_type_maps=True)
-    HinSAGENodeGenerator(G, schema=schema, batch_size=n_batch, num_samples=[1]).flow(
-        nodes_type_1
-    )
-
-    # Create manual schema without type maps
-    # Currently this raises an error
-    schema = G.create_graph_schema(create_type_maps=False)
-    with pytest.raises(RuntimeError):
-        HinSAGENodeGenerator(
-            G, schema=schema, batch_size=n_batch, num_samples=[1]
-        ).flow(nodes_type_1)
+    HinSAGENodeGenerator(
+        G, schema=schema, batch_size=n_batch, num_samples=[1], head_node_type="t1"
+    ).flow(nodes_type_1)
 
 
 def test_hinnodemapper_zero_samples():
@@ -636,9 +605,9 @@ def test_hinnodemapper_zero_samples():
     feature_sizes = {"t1": 1, "t2": 1}
     G, nodes_type_1, nodes_type_2 = example_hin_3(feature_sizes)
 
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[0, 0]).flow(
-        nodes_type_2
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[0, 0], head_node_type="t2"
+    ).flow(nodes_type_2)
 
     schema = G.create_graph_schema()
     sampling_adj = schema.type_adjacency_list(["t2"], 2)
@@ -655,9 +624,9 @@ def test_hinnodemapper_no_neighbors():
     feature_sizes = {"t1": 1, "t2": 1}
     G, nodes_type_1, nodes_type_2 = example_hin_3(feature_sizes)
 
-    mapper = HinSAGENodeGenerator(G, batch_size=batch_size, num_samples=[2, 1]).flow(
-        nodes_type_2
-    )
+    mapper = HinSAGENodeGenerator(
+        G, batch_size=batch_size, num_samples=[2, 1], head_node_type="t2"
+    ).flow(nodes_type_2)
 
     schema = G.create_graph_schema()
     sampling_adj = schema.type_adjacency_list(["t2"], 2)
@@ -747,8 +716,8 @@ def test_attri2vec_nodemapper_1():
     assert nf.shape == (1, n_feat)
 
     # This will fail as the nodes are not in the graph
-    with pytest.raises(KeyError):
-        Attri2VecNodeGenerator(G1, batch_size=2).flow(["A", "B"])
+    #    with pytest.raises(KeyError):
+    Attri2VecNodeGenerator(G1, batch_size=2).flow(["A", "B"])
 
 
 def test_attri2vec_nodemapper_2():
@@ -909,8 +878,7 @@ class Test_FullBatchNodeGenerator:
         )
         G = StellarGraph(G, node_type_name="node", node_features=node_features)
 
-        generator = FullBatchNodeGenerator(G, name="test", method=None)
-        assert generator.name == "test"
+        generator = FullBatchNodeGenerator(G, method=None)
         assert np.array_equal(feats, generator.features)
 
     def test_fullbatch_generator_init_3(self):